Added sampling of values for additional smoothing
The added sampling helps to remove bounces between some of the
steps.

Changed the threshold grouping to be a multiple/factor so it scales with
magnatude rather than being a fixed number.

Added a debug flag to get sample data and compare touched vs non-touched
pad values.

Moved the `ino` file into a folder with the same name

This makes it easier to open directly with Arduino IDE 2.x